requirements.txt
appnope==0.1.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
idna==2.6
ipdb==0.11
ipython==6.4.0
ipython-genutils==0.2.0
jedi==0.12.0
parso==0.2.1
pexpect==4.5.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.5.2
Pygments==2.2.0
PyYAML==3.12
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
urllib3==1.22
wcwidth==0.1.7
